break out favicon
