+++ components/common/page-layout/PageLayout.tsx
@@ -7,9 +7,9 @@ import { usePages } from 'hooks/usePages';
 import { useUser } from 'hooks/useUser';
 import Head from 'next/head';
 import * as React from 'react';
-import { getNonMacEmojiImage } from 'components/common/Emoji';
 import Header, { headerHeight } from './Header';
 import Sidebar from './Sidebar';
+import Favicon from './Favicon';
 
 const drawerWidth = 300;
 
@@ -75,26 +75,6 @@ const Drawer = styled(MuiDrawer, { shouldForwardProp: prop => prop !== 'open' })
 const HeaderSpacer = styled.div`
   min-height: ${headerHeight}px;
 `;
-
-function Favicon ({ icon }: { icon?: string | null }) {
-  const favicon = {
-    url: '/favicon.png',
-    type: 'image/png'
-  };
-  if (icon) {
-    const emojiImage = getNonMacEmojiImage(icon);
-    favicon.type = 'image/svg+xml';
-    if (emojiImage) {
-      favicon.url = emojiImage;
-    }
-    else {
-      favicon.url = `data:image/svg+xml,<svg xmlns=%22http://www.w3.org/2000/svg%22 viewBox=%2210 0 100 100%22><text y=%22.9em%22 font-size=%2290%22>${icon}</text></svg>`;
-    }
-  }
-  return (
-    <link rel='icon' type={favicon.type} href={favicon.url} />
-  );
-}
 
 export function PageContainer ({ children }: { children: React.ReactNode }) {
   return (
